Dev (#232)
* notes for tool build errors

* tool build fail handled without x option

* update go

* try cache clean

* dont use getopts for ci

* debug

* counterfeiter now uses go modules

* exhaustruct in rest package and notes

* custom job tracking for install-tools - pending xo update

* xo duplicate constraint fix for vertically partitioned tables

* better having clause notes

* update

* dynamic cache index up to date in post migrations and down migration init

* notes for migration notices

* diff check with untracked files

* refactor for better having clause

* notes

* notes for possible schema dump setup

* prevent trimming on go files locally

* notes for better schema dumping

* fix pg dump command for use in migration schema dump

* update

* save gen schema

* notes for actually useful migration comments

* comments in table example

* update gitignore

* set search path at conf level and fix pg format ignore

* dependabot

* remove dead m2o code - handled in o2o

* kinopenapi notes

* showcase nested objects and arrays in query params - must fix kin openapi

* fix cursors

* ci vet

* update runtime

* v2 for oapi codegen

* notes for pre build cmd

* notes for pre build cmd

* notes for pre build cmd

* update air

* update

* update

* update vendor

* update

* working on better xo join naming

notes

some broken updates

some broken updates

update

* clearer m2m joins

* add extra_schema m2m showcase

* update tests - pending m2m naming fix when lookup column name differs

* m2m prepends lookup join name when ref differs - pending extratypes sync

* update

* update

* m2m joins add ref prefix only when ref column differs - e.g. SellersJoin in books implicit book sellers

* update

* update tests

+++ bin/.helpers.sh
@@ -82,20 +82,40 @@ list_descendants() {
   echo "$desc_pids"
 }
 
-# waits for parallel processes to finish sucessfully, signalling SIGUSR1 otherwise.
+# Accepts flags:
+#    --no-kill    Do not immediately exit.
+# It does not store information the failed command. To keep track of all failures,
+# use:
+# 	for pid in "${pids[@]}"; do
+# 	  cmd=$(jobs -l | grep "$pid")
+# 	  wait -fn "$pid" || echo "Background job failed: $cmd"
+# 	done
 wait_without_error() {
   local -i err=0 werr=0
+  local kill=true
+
+  while [[ $# -gt 0 ]]; do # getopts fails in CI (non interactive)
+    case "$1" in
+    --no-kill) kill=false ;;
+    *) echo "Invalid option: $1" >&2 ;;
+    esac
+    shift
+  done
+
   while
-    wait -fn || werr=$? # do not quote
-    ((werr != 127))     # 127: not found
+    wait -fn || werr=$?
+    ((werr != 127))
   do
     err=$werr
-    ((err == 0)) || break # handle error as soon as it happens
+    ((err == 0)) || break # handle as soon as it happens
   done
-  #trap 'wait || :' EXIT # wait for all jobs before exiting (regardless of handling above)
+
   if ((err != 0)); then
+    sleep 0.2 # wait for all stdout/err
     echo "A job failed" >&2
-    kill -s SIGUSR1 $PROC
+    if $kill; then
+      kill -s SIGUSR1 $PROC
+    fi
     return 1
   fi
 }
@@ -378,7 +398,8 @@ show_tracebacks() {
 # Parameters:
 #   - Output .md5 file
 #   - Files or directories to cache
-#   - Optionally pass glob patterns to exclude.
+#   - Optionally pass glob patterns to exclude via --exclude [pattern].
+#   - Optionally pass --no-regen disable external cache invalidation.
 cache_all() {
   local excludes=()
   local args=()
